Open MotionCaptureWindow from MainWindow
